requirements.txt
google-generativeai
duckduckgo-search
jinja2
fastapi
uvicorn
python-multipart
python-dotenv
